Publish to maven central

+++ release.sh
@@ -19,14 +19,22 @@ fi
 echo "${RELEASE_TYPE}"
 
 forceNewVersion="false"
+snapshot="false"
 
+shift
 until [[ "$#" == "0" ]]; do
     case "$1" in
         --force-new-version )
             forceNewVersion="true"
             ;;
+        --snapshot )
+            snapshot="true"
+            ;;
+        * )
+          echo "Unknown option: $1"
+          exit 1
+          ;;
     esac
-
     shift
 done
 
@@ -36,27 +44,35 @@ echo "forceNewVersion=${forceNewVersion}"
 previousVersion=$(cat VERSION)
 alreadyVersioned=$(git tag -l --points-at HEAD "${previousVersion}")
 if [[ "${alreadyVersioned}" != "" ]]; then
-  if [[ "${forceNewVersion}" = "false" ]]; then
-      echo "Current HEAD already contains version tag ${alreadyVersioned}"
-      echo "Ignoring release command."
-      exit 1
+  echo "Current HEAD already contains version tag ${alreadyVersioned}"
+fi
+
+newVersion="false"
+if [[ "${alreadyVersioned}" = "" ]] || [[ "${forceNewVersion}" = "true" ]]; then
+  if [[ "${snapshot}" = "false" ]]; then
+    newVersion="true"
+
+    # bump (increment) version
+    docker run --rm -v "$PWD":/app treeder/bump "${RELEASE_TYPE}"
   fi
 fi
 
-# bump (increment) version
-docker run --rm -v "$PWD":/app treeder/bump "${RELEASE_TYPE}"
-version=`cat VERSION`
+version=$(cat VERSION)
 echo "Version: ${version}"
 
 # run build & tests
 ./gradlew clean build
 
-# tag it
-git add -A
-git commit -m "Release version ${version}"
-git tag -a "${version}" -m "Version ${version}"
-git push
-git push --tags
+if [[ "${newVersion}" = "true" ]]; then
+  # tag it
+  git add -A
+  git commit -m "Release version ${version}"
+  git tag -a "${version}" -m "Version ${version}"
+  git push
+  git push --tags
+fi
+
+export SNAPSHOT_RELEASE="${snapshot}"
 
 # publish it
 ./gradlew :lock-provider:publish
